moved button directive
